MySQLDialect.cs: Define power to return double to harmonize with other dialects (makes use of Math.Pow() in Linq queries work). NH-3381.

+++ src/NHibernate/Dialect/MySQLDialect.cs
@@ -140,7 +140,7 @@ namespace NHibernate.Dialect
 			
 			RegisterFunction("rand", new NoArgSQLFunction("rand", NHibernateUtil.Double));
 			
-			RegisterFunction("power", new StandardSQLFunction("power", NHibernateUtil.Single));
+			RegisterFunction("power", new StandardSQLFunction("power", NHibernateUtil.Double));
 			
 			RegisterFunction("stddev", new StandardSQLFunction("stddev", NHibernateUtil.Double));
 			RegisterFunction("variance", new StandardSQLFunction("variance", NHibernateUtil.Double));
